requirements.txt
torch
torchvision
timm
safetensors
accelerate
numpy
pandas
Pillow
scipy
tqdm
